% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uci_dataset.R
\name{get_uci_dataset}
\alias{get_uci_dataset}
\title{Get UCI Dataset}
\usage{
get_uci_dataset(id)
}
\arguments{
\item{id}{Character or numeric ID of the dataset.}
}
\value{
None. Downloads the file and saves it to the current directory.
}
\description{
Takes a dataset id and downloads a dataset from the UCI Machine Learning Repository.
}
\details{
To view datasets you can browse this spreadsheet:
https://docs.google.com/spreadsheets/d/1_Smac2zek_Nvq7kLjOCvHduWs4or-6QgqRvOJV52nuA/edit?usp=sharing

Placeholder: A function will be developed to see areas and links in RStudio
and a shorter url will be provided later
}
\examples{
get_uci_dataset("5")
}
