% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mlr_task.R
\name{make_mlr_task}
\alias{make_mlr_task}
\title{Creates an mlr task from a data list}
\usage{
make_mlr_task(data, ...)
}
\arguments{
\item{data}{A data list as returned by \code{read_ucimlr} with option \code{read_as = 'list'}}

\item{...}{Additional parameters to be passed to \code{mlr::makeRegrTask} or \code{mlr::makeClassifTask}}
}
\description{
Creates an mlr task from a data list
}
