% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ucimlr.R
\name{read_ucimlr}
\alias{read_ucimlr}
\title{Read a dataset from the UCI Machine Learning Repository into a data frame}
\usage{
read_ucimlr(dataset, col_names = TRUE, read_as = "list")
}
\arguments{
\item{dataset}{The name of the dataset that you want to read}

\item{col_names}{Should column names be included? Default is TRUE}

\item{read_as}{The format that should be returned. Default is \code{'df'}
\itemize{
  \item \code{'df'} Return a data frame
  \item \code{'tbl'} Return a tbl (see package \code{tibble})
  \item \code{'task'} Return a task (see package \code{mlr})
  \item \code{'list'} Return a data list containing elements \code{data} and \code{properties}
}}
}
\description{
Read a dataset from the UCI Machine Learning Repository into a data frame
}
\examples{
blog <- read_ucimlr("blog")
}
