% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_review_ratings.R
\docType{data}
\name{travel_review_ratings}
\alias{travel_review_ratings}
\title{Travel Review Ratings}
\format{A data frame with 5456 observations on the following 25 variables.
\enumerate{
  \item user id
  \item churches
  \item resorts
  \item beaches
  \item parks
  \item theatres
  \item museums
  \item malls
  \item zoo
  \item restaurants
  \item pubs/bars
  \item local services
  \item burger/pizza shops
  \item hotels/other lodgings
  \item juice bars
  \item art galleries
  \item dance clubs
  \item swimming pools
  \item gyms
  \item bakeries
  \item beauty & spas
  \item cafes
  \item view points
  \item monuments
  \item gardens
}}
\source{
Shini Renjith, shinirenjith '@' gmail.com
}
\usage{
travel_review_ratings
}
\description{
This data set is populated by capturing user ratings from Google reviews.
Reviews on attractions from 24 categories across Europe are considered.
Google user rating ranges from 1 to 5 and average user rating per category
is calculated.
}
\details{
Google reviews on attractions from 24 categories across Europe are
considered. Google user rating ranges from 1 to 5 and average user rating
per category is calculated.
}
\references{
Renjith, Shini, A. Sreekumar, and M. Jathavedan. 2018. Evaluation of
Partitioning Clustering Algorithms for Processing Social Media Data in
Tourism Domaina.

https://archive.ics.uci.edu/ml/machine-learning-databases/00485/

https://archive.ics.uci.edu/ml/datasets/Tarvel+Review+Ratings
}
\keyword{datasets}
