% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucinews.R
\name{ucinews}
\alias{ucinews}
\title{Latest UCI MLR News}
\format{
\enumerate{
  \item Date: The date information was reported.
  \item Info: News description.
}
}
\usage{
ucinews(quiet = FALSE)
}
\arguments{
\item{quiet}{If TRUE, a message informs the user where current UCI data is
coming from. If FALSE, no message is displayed. Defaults to FALSE.}
}
\description{
This function scrapes data from UCI's Machine Learning repository.
Specifically, it's scrapes a table from the main page. The table
contains latest information pertaining to the site.
}
\references{
\url{https://archive.ics.uci.edu/ml/index.php}
}
