% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucidata.R
\name{ucidata}
\alias{ucidata}
\title{Current UCI Datasets}
\format{
\enumerate{
  \item Name: The name of the dataset.
  \item Type: Multivariate, time-series, etc.
  \item Task: Classification, regression, etc.
  \item Variable Types: Categorical, integer, etc.
  \item Observations: Number of rows
  \item Variables: Number of columns
  \item Year: The year the dataset was donated
}
}
\usage{
ucidata(result = "raw", url = "https://archive.ics.uci.edu/ml/datasets.php")
}
\arguments{
\item{result}{One of raw, nest, or flat.}

\item{url}{The URL that points to the UCIMLR dataset table.}
}
\description{
This function scrapes data from UCI's Machine Learning repository.
Specifically, it's scrapes a table from the datasets page. The table
describes characteristics about the data.
}
\references{
\url{https://archive.ics.uci.edu/ml/datasets.html}
}
